requirements.txt
annotated-types==0.6.0
anyio==4.0.0
certifi==2023.7.22
deprecation==2.1.0
exceptiongroup==1.1.3
gotrue==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.26.1
packaging==23.2
pandas==2.1.2
postgrest==0.11.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.0
six==1.16.0
sniffio==1.3.0
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
typing_extensions==4.8.0
tzdata==2023.3
websockets==10.4
QuantLib
seaborn
plotly
numpy_financial
